+++ time/src/main/java/io/spine/time/validation/TimeOptionsProvider.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,10 +24,20 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-@CheckReturnValue
-@NullMarked
-package io.spine.time.validate.given;
+package io.spine.time.validation;
 
-import com.google.errorprone.annotations.CheckReturnValue;
+import com.google.auto.service.AutoService;
+import com.google.protobuf.ExtensionRegistry;
+import io.spine.option.OptionsProvider;
 
-import org.jspecify.annotations.NullMarked;
+/**
+ * Registers time-related proto options introduced by this library.
+ */
+@AutoService(OptionsProvider.class)
+public class TimeOptionsProvider implements OptionsProvider {
+
+    @Override
+    public void registerIn(ExtensionRegistry registry) {
+        TimeOptionsProto.registerAllExtensions(registry);
+    }
+}
