+++ time/src/test/proto/spine/time/validation/messages.proto
@@ -26,13 +26,13 @@
 
 syntax = "proto3";
 
-package spine.time.validate;
+package spine.time.validation;
 
 import "spine/options.proto";
 import "spine/time_options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.time.validate";
+option java_package = "io.spine.time.validation.given";
 option java_outer_classname = "MessagesProto";
 option java_multiple_files = true;
